refactor(VO-595): AgreementSection, TermsModal 컴포넌트 폴더 구조 정리
