+++ frontend/src/app/(auth)/signup/_components/TermsModal/TermsModal.tsx
